Cleaned up code in VolumeSlider.swift

+++ Utterance/SettingsOptions/VolumeSlider.swift
@@ -6,20 +6,14 @@
 //
 
 import SwiftUI
+import CoreHaptics
 
 struct VolumeSlider: View {
-    @State private var volumeLevel: Double = 100
+    @State private var volumeLevel: Float = 100
     
     var body: some View {
-        //        VStack {
-        //            Text("Volume: \(volumeLevel, specifier: "%.0f")")
-        //                .font(.subheadline.bold())
-        //                .foregroundColor(.white)
-        
         Slider(value: $volumeLevel, in: 0...100)
-        //        }
-        //        .background(Color.black)
-        //        .padding()
+        
     }
 }
